Cleanup dependencies, icon, and documentation
Removed a bunch of unused dependencies and files. Fixed the
documentation to actually contain useful information about how
the plugin works. Fixed the widget to contain all properties
and to use newer widget types.

Default the argument type to string and left out the fact that
type is supported from the documentation. This is because schema
is handled in a very custom way that does not match any schema
used anywhere else, arrays seem to be written specifically for
wrangler directives, and other issues.

Also fixed the icon to be a square so it does not get stretched.